파일 업로드 기능 추가

+++ sevenFS/src/main/java/kr/or/ddit/sevenfs/vo/webfolder/WebFolderFileVO.java
@@ -8,10 +8,9 @@ import java.util.List;
 @Data
 public class WebFolderFileVO {
     private int folderNo;
-    private int atchFileNo;
+    private long atchFileNo;
     private int fileDwldCo; // 파일 다운르도 카운트
     private String fileUploadEmpno;
 
     private AttachFileVO attachFileVO;
-
 }
